moved mbtu_m3 to economics
