+++ examples/CSP50compare.py
@@ -7,23 +7,21 @@ import matplotlib.pyplot as plt
 
 import numpy_financial as npf
 
+from CombiCSP import HOYS_DEFAULT
 import CombiCSP.SolarGeometry as sgh
 import CombiCSP.misc as cspm
 import CombiCSP.economics as cspe
-# from CombiCSP.CSP import *
-# from CombiCSP.CSP import  solarII, IAM_tow, di_sst, costhetai_NS \
-#     , IAM_tro , Ac,costhetai_EW, Cg_tro
+
 from CombiCSP.solar_tower import solarII,IAM_tow
 from CombiCSP.solar_trough import di_sst, IAM_tro, costhetai_NS, costhetai_EW, Ac, Cg_tro
 
 from CombiCSP.storage import Tr
 
 #import pcm
-# from CSPecon import *
 from CSPecon import csp_area_costs, power_block_cost, csp_energy_price \
     ,csp_discount_rate, oil_price, Eoil
 
-hoy = sgh.HOYS_DEFAULT
+hoy = HOYS_DEFAULT
 #%%
 
 # read data from local file
